+++ CHANGELOG.md
@@ -8,6 +8,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/), and this
 
 ## [Unreleased]
 
+# Added
+
+- Support for the OpenAI [`gpt-5.5`](https://developers.openai.com/api/docs/models/gpt-5.5) model.
+
 ### Changed
 
 - Update `nginx` to 1.30.3.